+++ react/OrderBy.js
@@ -6,6 +6,10 @@ import SelectionListOrderBy from './components/SelectionListOrderBy'
 
 export const SORT_OPTIONS = [
   {
+    value: '',
+    label: 'store/ordenation.default',
+  },
+  {
     value: 'OrderByScoreDESC',
     label: 'store/ordenation.relevance',
   },